# Switch to Ubuntu for faster pipelines