--- a/.github/workflows/fastlane.yml
+++ b/.github/workflows/fastlane.yml
@@ -36,8 +36,10 @@ jobs:
       - name: Grant execute permission for gradlew
         run: chmod u+x gradlew
 
-      - name: Install Fastlane
-        run: bundle install
+      - name: Install Ruby & Fastlane
+        uses: ruby/setup-ruby@v1
+        with:
+          bundler-cache: true # runs 'bundle install' and caches installed gems automatically
 
       - name: Run unit tests
         run: bundle exec fastlane android test
@@ -74,9 +76,11 @@ jobs:
       - name: Grant execute permission for gradlew
         run: chmod u+x gradlew
 
-#      - name: Install Fastlane
-#        run: bundle install
-
+#      - name: Install Ruby & Fastlane
+#        uses: ruby/setup-ruby@v1
+#        with:
+#          bundler-cache: true
+#
 #      - name: Run instrumented tests
 #        run: bundle exec fastlane android ui_test api:${{ matrix.api-level }}
       - name: Run instrumented tests