Copying the changes from https://github.com/Abhiek187/calculators/pull/128 because the performance bump is so good!

Let's compare the performance from my benchmarks here: https://github.com/Abhiek187/ez-recipes-android/pull/10#issuecomment-1304997300 (API 33)

| Step | Ubuntu | macOS |
| :-: | :-: | :-: |
| Change owner | 6s | 0s |
| Accept licenses | 3s | 6s |
| Install build/platform tools | 2s | 2s |
| Install emulator | 7s | 16s |
| Install system images | 2m 6s | 41s |
| Create AVD | 1s | 2s |
| Configure 2 cores | 0s | 0s |
| Start emulator | 10s | 10s |
| Wait to boot | 24s | 5m 59s |
| Disable animations | 1s | 1s |
| Run tests | 2m 17s | 7m 40s |
| Kill emulator | 0s | 0s |
| Total | 5m 22s | 15m 2s |

Booting up the emulator is 15x faster and running all the tests is 3x faster on Ubuntu than on macOS.